pyproject.toml
[project]
name = "git-grok"
version = "0.1.0"
description = "A simple Q&A agent that can answer questions about a GitHub repository"
readme = "README.md"
authors = [
    { name = "Nima Nooshiri", email = "nima.nooshiri@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.59.0",
    "pydantic-ai>=1.46.0",
    "python-frontmatter>=1.1.0",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "semantic-text-splitter>=0.29.0",
    "sentence-transformers>=5.2.0",
    "tokenizers>=0.22.2",
]

[project.scripts]
git-grok = "git_grok:main"

[build-system]
requires = ["uv_build>=0.9.25,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.13",
    "ty>=0.0.12",
]
